requirements.txt
mcp>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
lighthouse>=1.0.0
pydantic>=2.5.0
asyncio-throttle>=1.0.2
aiofiles>=23.2.1
requests>=2.31.0
lxml>=4.9.0
urllib3>=2.0.0
sqlite3
python-dotenv>=1.0.0
dataclasses-json>=0.6.0
typing-extensions>=4.8.0